+++ CHANGELOG.markdown
@@ -1,3 +1,4 @@
 # PhantomJS QUnit JUnit JSCoverage Cobertura
 
 * Prototype
+    * added optional argument --mockups to specify where your mockup files are located